# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Backend:Demo",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/userBasics/": {
      "post": {
        "tags": [
          "userBasic"
        ],
        "summary": "Add new user",
        "description": "",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_user"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/create_user"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "okk"
          }
        },
        "security": [
          {
            "api_key": {}
          }
        ]
      },
      "get": {
        "tags": [
          "userBasic"
        ],
        "summary": "Fetch All Users",
        "description": "",
        "operationId": "fetchAll",
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "okk"
          }
        },
        "security": [
          {
            "api_key": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "create_user": {
        "required": [
          "name",
          "place",
          "email",
          "phone",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "anyName"
          },
          "place": {
            "type": "string",
            "example": "place1"
          },
          "email": {
            "type": "string",
            "example": "ran@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "111"
          },
          "password": {
            "type": "string",
            "example": "123"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}